Removed all private uses of TF
