+++ kerastuner/abstractions/__init__.py
@@ -16,11 +16,5 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from .tensorflow import TENSORFLOW, TENSORFLOW_UTILS
 from .host import Host
 from .display import IS_NOTEBOOK
-
-from .io import Open, makedirs, exists, rmtree, glob, remove, copy
-from .io import write_file, read_file, create_directory, reload_model
-from .io import get_config_filename, get_weights_filename
-from .io import get_results_filename
