--- kerastuner/abstractions/io/__init__.py
@@ -1,23 +1,0 @@
-# Copyright 2019 The Keras Tuner Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-
-from .io import Open, makedirs, exists, rmtree, glob, remove, copy
-from .io import write_file, read_file, create_directory, reload_model
-from .io import get_config_filename, get_weights_filename
-from .io import get_results_filename
