+++ kerastuner/engine/tuner_utils.py
@@ -126,3 +126,11 @@ def average_histories(histories):
     # Convert {str: [float]} to [{str: float}]
     averaged = [dict(zip(metrics, vals)) for vals in zip(*averaged.values())]
     return averaged
+
+
+def maybe_compute_model_size(model):
+    """Compute the size of a given model, if it has been built."""
+    if model.built:
+        params = [keras.backend.count_params(p) for p in model.trainable_weights]
+        return int(np.sum(params))
+    return 0
